# docs/openapi.yaml
openapi: 3.1.0
info:
  title: REKAM API
  version: 0.1.0
  description: Backend API for REKAM local and on-premise deployment.
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ActivityInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: [string, "null"] }
        description_json: { type: [object, "null"] }
        event_date: { type: string, format: date }
        event_end_date: { type: [string, "null"], format: date }
        location: { type: [string, "null"] }
        use_sections: { type: boolean }
        team_id: { type: [string, "null"], format: uuid }
        district_id: { type: [string, "null"], format: uuid }
...
paths:
...
  /api/activities:
    get:
      summary: List activities
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: team_id, in: query, schema: { type: string, format: uuid } }
        - { name: district_id, in: query, schema: { type: string, format: uuid } }
        - { name: archived, in: query, schema: { type: boolean } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Paginated activities.
    post:
      summary: Create activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Activity created.
  /api/activities/{id}:
    get:
      summary: Activity detail with sections, media, and attachments
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Activity detail.
    put:
      summary: Update activity
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "200":
          description: Activity updated.
    delete:
      summary: Archive or hard delete activity
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: hard, in: query, schema: { type: boolean }, description: SUPER_ADMIN only. }
      responses:
        "200":
          description: Activity archived or deleted.
  /api/activities/{id}/versions:
    get:
      summary: List activity versions
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Version list.
  /api/activities/{id}/versions/{versionId}/restore:
    post:
      summary: Restore activity fields from a previous version
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: versionId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Version restored.
...